requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas==1.2.4
seaborn==0.11.1
numpy==1.20.1
matplotlib==3.4.1
scikit-learn==0.24.2
xgboost==1.4.2
pickle4==0.0.1
requests==2.25.1
scipy==1.6.2
statsmodels==0.12.2
tslearn==0.5.2
pip>=21.0.1
plotly==5.7.0
streamlit==1.9.0
streamlit-lottie==0.0.3
Pillow==8.2.0
protobuf~=3.19.0
hydralit_components==1.0.10
pyproj
